+++ fizjo/src/main/java/sopi/module/person/controller/UserControllerRest.java
@@ -11,12 +11,17 @@ import javax.validation.Valid;
 
 import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import sopi.module.auth.security.AuthUtils;
 import sopi.module.person.model.Profile;
 import sopi.module.person.model.ProfileModel;
 import sopi.module.person.model.Role;
@@ -32,6 +37,7 @@ public class UserControllerRest {
 	@Autowired UserModel userModel;
 	@Autowired RoleModel roleModel;
 	@Autowired ProfileModel profileModel;
+	@Autowired AuthUtils auth;
 	
 	@RequestMapping(value="*", method=RequestMethod.OPTIONS)
 	public boolean options() {
@@ -39,37 +45,62 @@ public class UserControllerRest {
 	}
 	
 	@RequestMapping(value="/get", method=RequestMethod.GET)
-	public List<User> list() {
-		return userModel.getUsers();
+	public ResponseEntity<?> list(HttpServletRequest request) {
+		List<User> list = new ArrayList<>();
+		
+		if (auth.checkRoles("ADMIN,USER")){
+			list.addAll(userModel.getUsers());
+		}
+		
+		if (list.isEmpty()){
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("ACCESS_DENIED");
+		} else {
+			return ResponseEntity.ok(list);
+		}
+		
 	}
 	
 	@RequestMapping(value="/get/{id}", method=RequestMethod.GET)
-	public User get(@PathVariable Long id) {
-		return userModel.getUser(id);
+	public ResponseEntity<?> get(@PathVariable Long id, HttpServletRequest request) {
+		
+		if (!auth.checkRoles("ADMIN,USER")){
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("ACCESS_DENIED");
+		}
+		
+		return ResponseEntity.ok(userModel.getUser(id));
 	}
 	
 	@RequestMapping(value="/getRoles/{id}", method=RequestMethod.GET)
-	public List<Role> getUserRoles(@PathVariable Long id) {
+	public List<Role> getUserRoles(@PathVariable Long id, HttpServletRequest request) {
 		return userModel.getUserRoles(id);
 	}
 	
 	@RequestMapping(value="/getRoles", method=RequestMethod.GET)
-	public List<Role> getRoles() {
+	public List<Role> getRoles(HttpServletRequest request) {
 		return roleModel.getRoles();
 	}
 	
 	@RequestMapping(value="/getAvailRoles/{id}", method=RequestMethod.GET)
-	public List<Role> getAvailRoles(@PathVariable Long id) {
+	public List<Role> getAvailRoles(@PathVariable Long id, HttpServletRequest request) {
 		List<Role> roles = roleModel.getRoles();
 		List<Role> userRoles = userModel.getUserRoles(id);
 		
 		roles.removeAll(userRoles);
 		
+		if (!auth.checkRoles("ADMIN")){
+			Role admin = roleModel.getRole("ADMIN");
+			roles.remove(admin);
+		}
+		
 		return roles;
 	}
 	
 	@RequestMapping(value="/setUserRole/{id}/{role}", method=RequestMethod.POST)
-	public StatusHelper setUserRole(@PathVariable Long id, @PathVariable String role) {
+	public ResponseEntity<?> setUserRole(@PathVariable Long id, @PathVariable String role, HttpServletRequest request) {
+		
+		if (!auth.checkRoles("ADMIN,USER")){
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("ACCESS_DENIED");
+		}		
 		
 		List<Object> param = new ArrayList<Object>();
 		param.add(id);
@@ -77,49 +108,62 @@ public class UserControllerRest {
 		
 		try {
 			userModel.setUserRole(id, role);
-			return new StatusHelper(true, "Rola została dodana", param);
+			return ResponseEntity.ok(new StatusHelper(true, "Rola została dodana", param));
 		} catch (Exception e) {
-			return new StatusHelper(false, "Wystąpił błąd: " + e.getMessage(), param);
+			return ResponseEntity.ok(new StatusHelper(false, "Wystąpił błąd: " + e.getMessage(), param));
 		}
 	}
 	
 	@RequestMapping(value="/deleteUserRole/{id}/{role}", method=RequestMethod.DELETE)
-	public StatusHelper deleteUserRole(@PathVariable Long id, @PathVariable String role){
+	public ResponseEntity<?> deleteUserRole(@PathVariable Long id, @PathVariable String role, HttpServletRequest request){
+		
+		if (!auth.checkRoles("ADMIN,USER")){
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("ACCESS_DENIED");
+		}	
+		
 		try {
 			boolean status = userModel.deleteUserRole(id, role);
-			return new StatusHelper(true, "Rola została usunięta.", status);
+			return ResponseEntity.ok(new StatusHelper(true, "Rola została usunięta.", status));
 		} catch (Exception e) {
-			return new StatusHelper(false, "Wystąpił błąd:" + e.getMessage(), null);
+			return ResponseEntity.ok(new StatusHelper(false, "Wystąpił błąd:" + e.getMessage(), null));
 		}
 	}
 	
 	@RequestMapping(value="/set/{id}", method=RequestMethod.PUT)
-	public StatusHelper set(@PathVariable Long id, @Valid @RequestBody User user) {
-		
-		if (id != user.getUserId()) {
-			return new StatusHelper(false, "Niezgodne numery Id! " + id.toString() + " != " + user.getUserId().toString(), user);
-		}
+	public ResponseEntity<?> set(@PathVariable Long id, @Valid @RequestBody User user, HttpServletRequest request) {
 		
 		User checkUser = userModel.getUser(user.getUsername());
 		
+		if (!auth.checkRoles("ADMIN,USER")){
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("ACCESS_DENIED");
+		}
+		
+		if (id != user.getUserId()) {
+			return ResponseEntity.ok(new StatusHelper(false, "Niezgodne numery Id! " + id.toString() + " != " + user.getUserId().toString(), user));
+		}
+				
 		if (checkUser != null && checkUser.getUserId() != user.getUserId()) {
-			return new StatusHelper(false, "Podana nazwa użytkownika jest niedostępna.", user);
+			return ResponseEntity.ok(new StatusHelper(false, "Podana nazwa użytkownika jest niedostępna.", user));
 		}
 		
 		try {
 			userModel.save(user);
-			return new StatusHelper(true, "Rekord został zaktualizowany.", user);
+			return ResponseEntity.ok(new StatusHelper(true, "Rekord został zaktualizowany.", user));
 		} catch (Exception e) {
-			return new StatusHelper(false, "Wystąpił błąd: " + e.getLocalizedMessage(), user);
+			return ResponseEntity.ok(new StatusHelper(false, "Wystąpił błąd: " + e.getLocalizedMessage(), user));
 		}
 		
 	}
 	
 	@RequestMapping(value="/add", method=RequestMethod.POST)
-	public StatusHelper add(@Valid @RequestBody User user) {
+	public ResponseEntity<?> add(@Valid @RequestBody User user, HttpServletRequest request) {
+		
+		if (!auth.checkRoles("ADMIN,USER")){
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("ACCESS_DENIED");
+		}
 		
 		if (userModel.getUser(user.getUsername()) != null){
-			return new StatusHelper(false, "Wystąpił błąd: Wybrana nazwa użytkownika jest niedostępna.",user);
+			return ResponseEntity.ok(new StatusHelper(false, "Wystąpił błąd: Wybrana nazwa użytkownika jest niedostępna.",user));
 		}
 		
 		if (userModel.getUser(user.getUserId()) == null) {
@@ -131,27 +175,41 @@ public class UserControllerRest {
 				user.setPassword("PASSWORD_TEST");
 				
 				userModel.saveNew(user);
-				return new StatusHelper(true, "Rekord został dodany.", user);
+				return ResponseEntity.ok(new StatusHelper(true, "Rekord został dodany.", user));
 			} catch (Exception e) {
-				return new StatusHelper(false, "Wystąpił błąd: " + e.getMessage(), user);
+				return ResponseEntity.ok(new StatusHelper(false, "Wystąpił błąd: " + e.getMessage(), user));
 			}
 		} else {
-			return new StatusHelper(false, "Wystąpił błąd: Konto użytkownika dla wybranego profilu już istnieje.",user);
+			return ResponseEntity.ok(new StatusHelper(false, "Wystąpił błąd: Konto użytkownika dla wybranego profilu już istnieje.",user));
 		}
 	}
 	
 	@RequestMapping(value="/getAvailProfiles", method=RequestMethod.GET)
-	public List<Profile> getProfiles() {
-		return profileModel.getProfileWithoutUser();
+	public ResponseEntity<?> getProfiles(HttpServletRequest request) {
+		
+		List<Profile> profiles = profileModel.getProfileWithoutUser();		
+		
+		if (auth.checkRoles("ADMIN,USER")){
+			return ResponseEntity.ok(profiles);
+		} else{
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("ACCESS_DENIED");
+		} 		
 	}
 	
 	@RequestMapping(value="/delete/{id}")
-	public StatusHelper deleteUser(@PathVariable Long id){
+	public ResponseEntity<?> deleteUser(@PathVariable Long id, HttpServletRequest request){
+		
+		User user = userModel.getUser(id);
+		
+		if (!auth.checkRoles("ADMIN,USER")){
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("ACCESS_DENIED");
+		}
+		
 		try {
 			userModel.deleteUser(id);
-			return new StatusHelper(true,"Użytkownik został usunięty", id);
+			return ResponseEntity.ok(new StatusHelper(true,"Użytkownik został usunięty", id));
 		} catch (Exception e) {
-			return new StatusHelper(false,"Wystąpił błąd: " + e.getMessage(), id);
+			return ResponseEntity.ok(new StatusHelper(false,"Wystąpił błąd: " + e.getMessage(), id));
 		}
 	}
 	
@@ -162,5 +220,20 @@ public class UserControllerRest {
 	    
 	    User user = userModel.getUserFromToken(authToken);
 	    return user;
+	}
+	@RequestMapping(value="/getLoggedRoles")
+	public ResponseEntity<?> getLoggedRoles(ServletRequest request, ServletResponse response){
+		HttpServletRequest httpRequest = (HttpServletRequest) request;
+	    String authToken = httpRequest.getHeader("X-Auth-Token");
+	    
+	    User user = userModel.getUserFromToken(authToken);
+	    
+	    List<String> roles = new ArrayList<>();
+	    
+	    for (GrantedAuthority role : SecurityContextHolder.getContext().getAuthentication().getAuthorities()){
+	    	roles.add(role.getAuthority());
+	    }
+	    	    
+	    return ResponseEntity.ok(roles);
 	}
 }
